+++ docs/changelog.rst
@@ -11,10 +11,11 @@ v0.1 (not yet released)
   `#1 <https://github.com/lsils/mockturtle/pull/1>`_
   `#4 <https://github.com/lsils/mockturtle/pull/4>`_
 
-* Algorithms: `cut_enumeration`, `lut_mapping`, `akers_synthesis`
+* Algorithms: `cut_enumeration`, `lut_mapping`, `akers_synthesis`, `collapse_mapped_network`
   `#2 <https://github.com/lsils/mockturtle/pull/2>`_
   `#7 <https://github.com/lsils/mockturtle/pull/7>`_
   `#9 <https://github.com/lsils/mockturtle/pull/9>`_
+  `#12 <https://github.com/lsils/mockturtle/pull/12>`_
 
 * Views: `topo_view`, `immutable_view`
   `#3 <https://github.com/lsils/mockturtle/pull/3>`_